delete event

+++ frontend/src/pages/DeleteEventPage.js
@@ -1,26 +1,30 @@
-// DeleteEventPage.js
-import React, { useEffect } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { useDispatch } from 'react-redux';
-import { useParams, useHistory } from 'react-router-dom';
+import { useParams, useNavigate } from 'react-router-dom';
 import { deleteEvent } from '../actions/eventActions';
 
 const DeleteEventPage = () => {
   const { id } = useParams();
   const dispatch = useDispatch();
-  const history = useHistory();
+  const navigate = useNavigate();
+  const hasConfirmed = useRef(false);
 
   useEffect(() => {
-    const confirmDelete = window.confirm('Are you sure you want to delete this event?');
-    if (confirmDelete) {
-      dispatch(deleteEvent(id)).then(() => {
-        history.push('/');
-      }).catch(error => {
-        console.error('Error deleting event:', error);
-      });
-    } else {
-      history.push(`/event/${id}`);
+    if (!hasConfirmed.current) {
+      const confirmDelete = window.confirm('Are you sure you want to delete this event?');
+      hasConfirmed.current = true;
+
+      if (confirmDelete) {
+        dispatch(deleteEvent(id)).then(() => {
+          navigate('/events');
+        }).catch(error => {
+          console.error('Error deleting event:', error);
+        });
+      } else {
+        navigate(`/event/${id}`);
+      }
     }
-  }, [dispatch, id, history]);
+  }, [dispatch, id, navigate]);
 
   return (
     <div>
